add 7702 check to parrallel
